Adds Completed Specio Pipeline
The pipeline is now usable via the local docker environment. SFTP is
still pending. The pipeline writes a .lock file to indicate that is is
running, and removes it when complete.

Pipelines can be configured using the sample .yml file in the
customers/codebiosys directory and more documentation is coming.

This commit greatly increases the complexity of Specio as it adds a
celery dependency, selenium, and a redis queue.