Adds Automatic Video Recordings and Starter Production Environment
Multiple new fixes and features have been added.

Video Recordings
----------------

A new set of tasks have been added to the pipeline to facilitate the
automatic recording of the running selenium tests.

In order to allow this, a new Flask app has been created inside of the
selenium container that accepts incoming recording start/stop requests.
This app uses FFmpeg to capture the X11 output from the selenium
container and dump it to a file. The flask app is managed via
supervisord in an effort to remain consistent with the selenium docker
project it runs inside of.

Due to the fact that the recordings must be asyncronous, the FFmpeg
recording is run inside of a deamonized tmux session. Once the tests
have finished, the tmux session id can be used to stop the recording and
the resulting mpeg-video can be copied to the customer's directory as
preferred.

In addition, a few startup scripts have been modified to reflect options
required by FFmpeg's X11Grab device. Notably Xvfb must listen for
incoming tcp connections.

Production Env
--------------

A new production compose file has been added to the project that boots
the project in a production-safe mode including disabling VNC for the
selenium container and activating the Specio/Worker production settings.

More work needs to be done here, but the initial version works.

Misc Fixes
----------

In addition to the two major additions listed above, the following
tweaks have been added.

- A new `recording_location` param is required in all run configuration
YAML files.
- The VNC password is now configurable rather than being hardcoded.
- Fixed a typo with bufffered Python output settings.
- All celery tasks now take a uniform `kwargs` variable in an effort to
remove the function signature confusion that can happen with reordering
and refactoring tasks.
  - Tasks now must accept and return a `kwargs` param.
  - Any data that they wish to pass on must be added to the `kwargs`
before returning.
- Backports production changes for internal routing.