Fixes development environment and adds new --force option.
There's a new force option that will ignore the presence of any existing
lockfile and perform the test run anyway.

This also fixes issues with the fact that the developement versions of
the selenium container did not boot the flask app.

Fixes an issue with logging in celery. Tasks awaiting chains don't log
properly so the `pipeline` task has been downgraded to a function.