Adds new logging configs, CLI, and run modes.
Specio has a new runmode where it can idle and wait for commands. This
is now the default mode.

To interact with the new idle process, either use the Docker (vanilla)
commands to exec into the running container, or use the new specio-cli
script to quickly execute commands in the specio head node.

Logging has also been fixed and should more reliably print to the kibana
logs now.